+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Openshift Serverless v1.3.0
 
+- Update Knative Serving to v0.10.0. See
+  [upstream release notes](https://github.com/knative/serving/releases/tag/v0.10.0)
+  for more information.
+- HTTPS based routes are now available for all Knative Services.
 - Fixed a bug where Routes have non-correct cross-namespaced OwnerReferences.